Shift int to uint32 / uint64, encode actual bytes

+++ src/util/constants.go
@@ -9,7 +9,7 @@ type ConstantsType struct {
 	Listen               bool          `json:"listen"`
 	MinPeers             int           `json:"maxPeers"`
 	MaxPeers             int           `json:"minPeers"`
-	MaxVSize             int           `json:"maxVSize"`
+	MaxVSize             uint64        `json:"maxVSize"`
 	PeerPingFreq         time.Duration `json:"peerPingFreq"`
 	PrintPeersUpdateFreq time.Duration `json:"printPeersUpdateFreq"`
 	RuntimeID            string        `json:"runtimeID"`
